# OPENAPI.yaml
openapi: 3.1.0
info:
  title: WorkMesh API
  version: "1.0"
  description: WorkMesh server 1.0.0 REST API 1.0. Non-stable capabilities are deployment-gated and default disabled.
servers:
  - url: http://localhost:3001
security:
  - SessionCookie: []
paths:
...
  /api/v1/projects:
    get:
      {
        x-workmesh-policy-id: route.listProjects,
        x-workmesh-actor-kinds: [human, agent],
        x-workmesh-feature-key: none,
        x-workmesh-feature-tier: stable,
        security: [{ SessionCookie: [] }, { AgentSessionToken: [] }],
        operationId: listProjects,
        parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }],
        responses:
          {
            "200": { $ref: "#/components/responses/Projects" },
            "401": { $ref: "#/components/responses/Unauthorized" },
          },
      }
    post:
      {
        x-workmesh-policy-id: route.createProject,
        x-workmesh-actor-kinds: [human, agent],
        x-workmesh-feature-key: none,
        x-workmesh-feature-tier: stable,
        security: [{ SessionCookie: [] }, { AgentSessionToken: [] }],
        operationId: createProject,
        parameters:
          [
            { $ref: "#/components/parameters/IdempotencyKey" },
            { $ref: "#/components/parameters/CsrfToken" },
          ],
        requestBody: { $ref: "#/components/requestBodies/ProjectInput" },
        responses:
          {
            "200": { $ref: "#/components/responses/Command" },
            "400": { $ref: "#/components/responses/BadRequest" },
            "401": { $ref: "#/components/responses/Unauthorized" },
            "403": { $ref: "#/components/responses/Forbidden" },
            "404": { $ref: "#/components/responses/NotFound" },
            "409": { $ref: "#/components/responses/Conflict" },
          },
      }
...
  /api/v1/work-items:
    get:
      {
        x-workmesh-policy-id: route.listWorkItems,
        x-workmesh-actor-kinds: [human, agent],
        x-workmesh-feature-key: none,
        x-workmesh-feature-tier: stable,
        security: [{ SessionCookie: [] }, { AgentSessionToken: [] }],
        operationId: listWorkItems,
        parameters:
          [
            { $ref: "#/components/parameters/Cursor" },
            { $ref: "#/components/parameters/Limit" },
            { $ref: "#/components/parameters/TeamIdQuery" },
            { $ref: "#/components/parameters/StatusIdQuery" },
            { $ref: "#/components/parameters/ProjectIdQuery" },
            { $ref: "#/components/parameters/PriorityQuery" },
            { $ref: "#/components/parameters/StatusCategoryQuery" },
            { $ref: "#/components/parameters/ResponsibleHumanActorIdQuery" },
            { $ref: "#/components/parameters/OwnerIdQuery" },
            { $ref: "#/components/parameters/MineQuery" },
            { $ref: "#/components/parameters/LabelQuery" },
            { $ref: "#/components/parameters/SearchQuery" },
          ],
        responses:
          {
            "200": { $ref: "#/components/responses/WorkItems" },
            "400": { $ref: "#/components/responses/BadRequest" },
            "401": { $ref: "#/components/responses/Unauthorized" },
          },
      }
    post:
      {
        x-workmesh-policy-id: route.createWorkItem,
        x-workmesh-actor-kinds: [human, agent],
        x-workmesh-feature-key: none,
        x-workmesh-feature-tier: stable,
        security: [{ SessionCookie: [] }, { AgentSessionToken: [] }],
        operationId: createWorkItem,
        parameters:
          [
            { $ref: "#/components/parameters/IdempotencyKey" },
            { $ref: "#/components/parameters/CsrfToken" },
          ],
        requestBody: { $ref: "#/components/requestBodies/WorkItemInput" },
        responses:
          {
            "200": { $ref: "#/components/responses/Command" },
            "400": { $ref: "#/components/responses/BadRequest" },
            "401": { $ref: "#/components/responses/Unauthorized" },
            "403": { $ref: "#/components/responses/Forbidden" },
            "404": { $ref: "#/components/responses/NotFound" },
            "409": { $ref: "#/components/responses/Conflict" },
          },
      }
...
  /api/v1/agent-sessions/{id}/ack:
    post:
      {
        x-workmesh-policy-id: route.acknowledgeAgentSession,
        x-workmesh-actor-kinds: [human, agent],
        x-workmesh-feature-key: none,
        x-workmesh-feature-tier: stable,
        security: [{ SessionCookie: [] }, { AgentSessionToken: [] }],
        operationId: acknowledgeAgentSession,
        parameters:
          [
            { $ref: "#/components/parameters/Id" },
            { $ref: "#/components/parameters/IdempotencyKey" },
            { $ref: "#/components/parameters/CorrelationId" },
          ],
        requestBody:
          { $ref: "#/components/requestBodies/AcknowledgeAgentSessionInput" },
        responses:
          {
            "200": { $ref: "#/components/responses/AgentSession" },
            "409": { $ref: "#/components/responses/Conflict" },
          },
      }
  /api/v1/agent-sessions/{id}/heartbeat:
    post:
      {
        x-workmesh-policy-id: route.heartbeatAgentSession,
        x-workmesh-actor-kinds: [human, agent],
        x-workmesh-feature-key: none,
        x-workmesh-feature-tier: stable,
        security: [{ SessionCookie: [] }, { AgentSessionToken: [] }],
        operationId: heartbeatAgentSession,
        parameters:
          [
            { $ref: "#/components/parameters/Id" },
            { $ref: "#/components/parameters/IdempotencyKey" },
            { $ref: "#/components/parameters/CorrelationId" },
          ],
        requestBody: { $ref: "#/components/requestBodies/HeartbeatInput" },
        responses:
          {
            "200": { $ref: "#/components/responses/AgentSession" },
            "409": { $ref: "#/components/responses/Conflict" },
          },
      }
...
  /api/v1/agent-sessions/{id}/stop-ack:
    post:
      {
        x-workmesh-policy-id: route.acknowledgeAgentSessionStop,
        x-workmesh-actor-kinds: [human, agent],
        x-workmesh-feature-key: none,
        x-workmesh-feature-tier: stable,
        security: [{ SessionCookie: [] }, { AgentSessionToken: [] }],
        operationId: acknowledgeAgentSessionStop,
        parameters:
          [
            { $ref: "#/components/parameters/Id" },
            { $ref: "#/components/parameters/IdempotencyKey" },
            { $ref: "#/components/parameters/IfMatch" },
            { $ref: "#/components/parameters/CorrelationId" },
          ],
        requestBody:
          { $ref: "#/components/requestBodies/StopAcknowledgementInput" },
        responses:
          {
            "200": { $ref: "#/components/responses/AgentSession" },
            "409": { $ref: "#/components/responses/Conflict" },
          },
      }
...
  /api/v1/inbox/{id}/claim:
    post: { x-workmesh-policy-id: route.claimInboxItem, x-workmesh-actor-kinds: [agent], x-workmesh-feature-key: none, x-workmesh-feature-tier: stable, security: [{ AgentSessionToken: [] }], operationId: claimInboxItem, parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/IdempotencyKey" }, { $ref: "#/components/parameters/CorrelationId" }], requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false } } } }, responses: { "200": { $ref: "#/components/responses/Json" }, "404": { $ref: "#/components/responses/NotFound" }, "409": { $ref: "#/components/responses/Conflict" } } }
...
  /api/v1/leases/{id}/heartbeat:
    post: { x-workmesh-policy-id: route.heartbeatLease, x-workmesh-actor-kinds: [human, agent], x-workmesh-feature-key: none, x-workmesh-feature-tier: stable, security: [{ SessionCookie: [] }, { AgentSessionToken: [] }], operationId: heartbeatLease, parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/IdempotencyKey" }], responses: { "200": { $ref: "#/components/responses/Json" } } }
...
components:
  securitySchemes:
    BootstrapToken:
      {
        type: apiKey,
        in: header,
        name: X-WorkMesh-Bootstrap-Token,
        description: "One-time deployment bootstrap credential. It is required for Workspace installation and is not a human or Agent session.",
      }
    SessionCookie:
      {
        type: apiKey,
        in: cookie,
        name: workmesh_session,
        description: "HttpOnly server-side session cookie.",
      }
    AgentSessionToken:
      {
        type: http,
        scheme: bearer,
        bearerFormat: WorkMeshSessionToken,
        description: "Short-lived agent session token; revocation is checked server-side.",
      }
    AgentInstallationToken:
      {
        type: http,
        scheme: bearer,
        bearerFormat: WorkMeshInstallationToken,
        description: "Installation-scoped bearer credential used by installation_target operations, including token exchange, refresh, and exact-target handoff actions. It cannot perform ordinary session work.",
      }
    GitHubWebhookSignature:
      {
        type: apiKey,
        in: header,
        name: X-Hub-Signature-256,
        description: "GitHub HMAC-SHA256 over the raw request body. X-GitHub-Delivery and X-GitHub-Event are also required and the server enforces its replay and delivery rules.",
      }
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Versioned, signed, opaque continuation cursor. It is bound to the route, workspace, actor, effective filters, and sort tuple; clients must not decode or modify it.
      schema: { type: string, minLength: 1, maxLength: 8192 }
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of authorized items returned in this page.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    IdempotencyKey:
      {
        name: Idempotency-Key,
        in: header,
        required: true,
        description: "Identifies one logical mutation attempt. Authentication and secret-bearing session mutations retain an encrypted replay for 15 minutes and protect key conflicts for 24 hours. A different operation, canonical request, or client context returns IDEMPOTENCY_KEY_REUSED; an exact retry after the replay window returns IDEMPOTENCY_REPLAY_EXPIRED and must use a new key.",
        schema: { type: string, minLength: 1 },
      }
    CsrfToken:
      {
        name: X-CSRF-Token,
        in: header,
        required: true,
        schema: { type: string, minLength: 1 },
      }
...
    TeamIdQuery:
      { name: teamId, in: query, schema: { type: string, format: uuid } }
    StatusIdQuery:
      { name: statusId, in: query, schema: { type: string, format: uuid } }
    ProjectIdQuery:
      { name: projectId, in: query, schema: { type: string, format: uuid } }
    PriorityQuery:
      {
        name: priority,
        in: query,
        schema: { $ref: "#/components/schemas/Priority" },
      }
    StatusCategoryQuery:
      {
        name: statusCategory,
        in: query,
        schema: { $ref: "#/components/schemas/StatusCategory" },
      }
    ResponsibleHumanActorIdQuery:
      {
        name: responsibleHumanActorId,
        in: query,
        schema: { type: string, format: uuid },
      }
    OwnerIdQuery:
      {
        name: ownerId,
        in: query,
        deprecated: true,
        schema: { type: string, format: uuid },
      }
    MineQuery: { name: mine, in: query, schema: { type: boolean } }
...
    SearchQuery:
      { name: search, in: query, schema: { type: string, minLength: 1 } }
  requestBodies:
...
    ProjectInput:
      {
        required: true,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ProjectInput" } },
          },
      }
...
  responses:
...
    Projects:
      {
        description: Projects,
        content:
          {
            application/json:
              {
                schema:
                  { type: object, additionalProperties: false, required: [items, nextCursor], properties: { items: { type: array, maxItems: 200, items: { $ref: "#/components/schemas/Project" } }, nextCursor: { type: [string, "null"] } } },
              },
          },
      }
...
    Command:
      {
        description: Mutation result,
        headers:
          {
            ETag:
              {
                schema: { type: string },
                description: Revision ETag when the result includes revision,
              },
          },
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Command" } },
          },
      }
    BadRequest:
      {
        description: Invalid input or request precondition,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
    Unauthorized:
      {
        description: Missing,
        expired,
        or invalid session,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
...
    Forbidden:
      {
        description: CSRF or authorization failure,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
...
    NotFound:
      {
        description: Resource does not exist in the current workspace,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
    Conflict:
      {
        description: "Revision or idempotency conflict. Authentication mutations may return IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_REPLAY_EXPIRED, or IDEMPOTENCY_REPLAY_UNAVAILABLE.",
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
...